Add support for browserify with polyfill generated from github/fetch.
