+++ package.json
@@ -1,9 +1,10 @@
 {
   "name": "node-fetch",
-  "version": "1.0.4",
+  "version": "1.1.0",
   "description": "A light-weight module that brings window.fetch to node.js and io.js",
   "main": "index.js",
   "scripts": {
+    "install": "node ./install",
     "test": "mocha test/test.js",
     "report": "istanbul cover _mocha -- -R spec test/test.js",
     "coverage": "istanbul cover _mocha --report lcovonly -- -R spec test/test.js && cat ./coverage/lcov.info | coveralls"
@@ -34,6 +35,7 @@
     "resumer": "0.0.0"
   },
   "dependencies": {
-    "encoding": "^0.1.11"
+    "encoding": "^0.1.11",
+    "whatwg-fetch": "^0.7.0"
   }
 }
